D18 : part 1,2

+++ src/main/java/fr/ela/aoc2023/D17.java
@@ -72,7 +72,7 @@ public class D17 extends AoC {
                     case SOUTH -> 'v';
                     case NORTH -> '^';
                 };
-                lines.get(n.position.y())[n.position.x()] = c;
+                lines.get((int) n.position.y())[(int) n.position.x()] = c;
             });
             return lines.stream().map(String::new).collect(Collectors.joining("\n"));
         }
